fix(theme): save theme in local storage after theme changing
